requirements.txt
yfinance>=0.2.31
fredapi>=0.5.1
pycoingecko>=3.1.0
plotly>=5.18.0
streamlit>=1.29.0
xgboost>=2.0.0
hmmlearn>=0.3.0
ta>=0.11.0
arch>=6.2.0
praw>=7.7.1
newsapi-python>=0.2.7
gdeltdoc>=1.4
vaderSentiment>=3.3.2
textblob>=0.17.1
feedparser>=6.0.10
apscheduler>=3.10.4
joblib>=1.3.2
pyarrow>=14.0.1
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
requests>=2.31.0
